Rename package Noisers to Noiser. Drop unused Bazel deps.
