+++ src/main/kotlin/org/wfanet/measurement/eventdataprovider/privacybudgetmanagement/BUILD.bazel
@@ -10,11 +10,7 @@ kt_jvm_library(
     srcs = glob(["*.kt"]),
     deps = [
         "//src/main/kotlin/org/wfanet/measurement/eventdataprovider/eventfiltration:event_filters",
-        "//src/main/kotlin/org/wfanet/measurement/eventdataprovider/noisers",
-        "//src/main/proto/wfa/measurement/api/v2alpha:differential_privacy_kt_jvm_proto",
-        "//src/main/proto/wfa/measurement/api/v2alpha:measurement_spec_kt_jvm_proto",
-        "//src/main/proto/wfa/measurement/api/v2alpha:protocol_config_kt_jvm_proto",
-        "//src/main/proto/wfa/measurement/api/v2alpha:requisition_spec_kt_jvm_proto",
+        "//src/main/kotlin/org/wfanet/measurement/eventdataprovider/noiser",
         "@wfa_common_jvm//imports/java/org/apache/commons:math3",
         "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/common",
     ],
