+++ src/test/kotlin/org/wfanet/measurement/eventdataprovider/noiser/BUILD.bazel
@@ -1,0 +1,24 @@
+load("@io_bazel_rules_kotlin//kotlin:jvm.bzl", "kt_jvm_test")
+
+kt_jvm_test(
+    name = "LaplaceNoiserTest",
+    srcs = ["LaplaceNoiserTest.kt"],
+    test_class = "org.wfanet.measurement.eventdataprovider.noiser.LaplaceNoiserTest",
+    deps = [
+        "//src/main/kotlin/org/wfanet/measurement/eventdataprovider/noiser",
+        "@wfa_common_jvm//imports/java/com/google/common/truth",
+        "@wfa_common_jvm//imports/java/org/junit",
+    ],
+)
+
+kt_jvm_test(
+    name = "GaussianNoiserTest",
+    srcs = ["GaussianNoiserTest.kt"],
+    test_class = "org.wfanet.measurement.eventdataprovider.noiser.GaussianNoiserTest",
+    deps = [
+        "//src/main/kotlin/org/wfanet/measurement/eventdataprovider/noiser",
+        "@wfa_common_jvm//imports/java/com/google/common/truth",
+        "@wfa_common_jvm//imports/java/org/junit",
+        "@wfa_common_jvm//imports/kotlin/kotlin/test",
+    ],
+)
