--- src/test/kotlin/org/wfanet/measurement/eventdataprovider/noisers/BUILD.bazel
@@ -1,28 +1,0 @@
-load("@io_bazel_rules_kotlin//kotlin:jvm.bzl", "kt_jvm_test")
-
-kt_jvm_test(
-    name = "LaplaceNoiserTest",
-    srcs = ["LaplaceNoiserTest.kt"],
-    test_class = "org.wfanet.measurement.eventdataprovider.noisers.LaplaceNoiserTest",
-    deps = [
-        "//src/main/kotlin/org/wfanet/measurement/eventdataprovider/noisers",
-        "//src/main/proto/wfa/measurement/api/v2alpha:measurement_spec_kt_jvm_proto",
-        "@wfa_common_jvm//imports/java/com/google/common/truth",
-        "@wfa_common_jvm//imports/java/com/google/common/truth/extensions/proto",
-        "@wfa_common_jvm//imports/java/org/junit",
-    ],
-)
-
-kt_jvm_test(
-    name = "GaussianNoiserTest",
-    srcs = ["GaussianNoiserTest.kt"],
-    test_class = "org.wfanet.measurement.eventdataprovider.noisers.GaussianNoiserTest",
-    deps = [
-        "//src/main/kotlin/org/wfanet/measurement/eventdataprovider/noisers",
-        "//src/main/proto/wfa/measurement/api/v2alpha:measurement_spec_kt_jvm_proto",
-        "@wfa_common_jvm//imports/java/com/google/common/truth",
-        "@wfa_common_jvm//imports/java/com/google/common/truth/extensions/proto",
-        "@wfa_common_jvm//imports/java/org/junit",
-        "@wfa_common_jvm//imports/kotlin/kotlin/test",
-    ],
-)
