+++ src/test/kotlin/org/wfanet/measurement/eventdataprovider/privacybudgetmanagement/BUILD.bazel
@@ -103,7 +103,6 @@ kt_jvm_test(
     test_class = "org.wfanet.measurement.eventdataprovider.privacybudgetmanagement.AcdpParamsConverterTest",
     deps = [
         "//src/main/kotlin/org/wfanet/measurement/eventdataprovider/privacybudgetmanagement:privacy_budget_manager",
-        "//src/main/proto/wfa/measurement/api/v2alpha:differential_privacy_kt_jvm_proto",
         "@wfa_common_jvm//imports/java/com/google/common/truth",
         "@wfa_common_jvm//imports/java/org/junit",
         "@wfa_common_jvm//imports/kotlin/kotlin/test",
